feat(toolkit): define optional keys for getuserinfo
