+++ packages/core/src/libraries/social.ts
@@ -1,4 +1,5 @@
-import type { GetSession } from '@logto/connector-kit';
+import type { GetSession, SocialUserInfo } from '@logto/connector-kit';
+import { socialUserInfoGuard } from '@logto/connector-kit';
 import type { User } from '@logto/schemas';
 import { ConnectorType } from '@logto/schemas';
 import type { Nullable } from '@silverhand/essentials';
@@ -6,8 +7,6 @@ import type { InteractionResults } from 'oidc-provider';
 import { z } from 'zod';
 
 import { getLogtoConnectorById } from '#src/connectors/index.js';
-import type { SocialUserInfo } from '#src/connectors/types.js';
-import { socialUserInfoGuard } from '#src/connectors/types.js';
 import RequestError from '#src/errors/RequestError/index.js';
 import { findUserByEmail, findUserByPhone } from '#src/queries/user.js';
 import assertThat from '#src/utils/assert-that.js';
