+++ packages/core/src/routes/interaction/utils/social-verification.ts
@@ -1,10 +1,9 @@
-import type { ConnectorSession } from '@logto/connector-kit';
+import type { ConnectorSession, SocialUserInfo } from '@logto/connector-kit';
 import type { SocialConnectorPayload } from '@logto/schemas';
 import { ConnectorType } from '@logto/schemas';
 import type { Provider } from 'oidc-provider';
 
 import { getLogtoConnectorById } from '#src/connectors/index.js';
-import type { SocialUserInfo } from '#src/connectors/types.js';
 import { getUserInfoByAuthCode } from '#src/libraries/social.js';
 import type { WithLogContext } from '#src/middleware/koa-audit-log.js';
 import {
